requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
alembic==1.17.1
amqp==5.3.1
annotated-types==0.7.0
anyio==4.11.0
APScheduler==3.11.0
asyncpg==0.30.0
attrs==25.4.0
bidict==0.23.1
billiard==4.2.2
catboost==1.2.8
celery==5.5.3
certifi==2025.10.5
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudpickle==3.1.2
colorama==0.4.6
colorlog==6.10.1
contourpy==1.3.3
Counter==1.0.0
cycler==0.12.1
fastapi==0.119.1
fastapi-socketio==0.0.10
fonttools==4.60.1
frozenlist==1.8.0
graphviz==0.21
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
joblib==1.5.2
kiwisolver==1.4.9
kombu==5.5.4
lightgbm==4.6.0
llvmlite==0.45.1
Mako==1.3.10
MarkupSafe==3.0.3
matplotlib==3.10.7
multidict==6.7.0
narwhals==2.12.0
numba==0.62.1
numpy==2.3.4
optuna==4.5.0
packaging==25.0
pandas==2.3.3
pillow==12.0.0
plotly==6.5.0
prompt_toolkit==3.0.52
propcache==0.4.1
psycopg2-binary==2.9.11
pydantic==2.12.3
pydantic_core==2.41.4
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-engineio==4.12.3
python-socketio==5.14.2
pytz==2025.2
PyYAML==6.0.3
redis==5.2.1
scikit-learn==1.7.2
scipy==1.16.2
shap==0.50.0
simple-websocket==1.1.0
six==1.17.0
slicer==0.0.8
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.48.0
threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uvicorn==0.38.0
vine==5.1.0
wcwidth==0.2.14
wsproto==1.2.0
xgboost==3.1.1
yarl==1.22.0
